nfpm.native_libs: move search_for_sonames rule into deb package
