+++ src/python/pants/backend/nfpm/native_libs/BUILD
@@ -1,9 +1,7 @@
 # Copyright 2025 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-python_sources(
-    overrides={"rules.py": dict(dependencies=["./deb"])},
-)
+python_sources()
 
 python_tests(
     name="tests",
