+++ src/python/pants/backend/nfpm/native_libs/deb/BUILD
@@ -1,7 +1,9 @@
 # Copyright 2025 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-python_sources()
+python_sources(
+    overrides={"rules.py": dict(dependencies=["./search_for_sonames.py"])},
+)
 
 python_tests(
     name="tests",
